requirements.txt
opencv-python==4.8.1.78
numpy==1.24.4
transformers==4.31.0
Pillow==9.4.0
torch==2.0.1
torchvision==0.15.2
moviepy==1.0.2
matplotlib==3.7.2